requirements.txt
PyMySQL==0.9.3
scipy==1.5.0
plotly==4.5.0
pytz==2019.3
pandas==1.0.1
numpy==1.18.1
TA_Lib==0.4.17
requests==2.22.0
python_dateutil==2.8.2
